Move inferrers to prim
